Added logout and account button
Before, clicking the button will logout the user, now there are two
buttons that allow the user to either logout or go to the account
page.
This marks the end of the nav bar and the people app. Hopefully
nothing will need to be added, not I will start on forum app.